requirements.txt
asgiref==3.4.1
Django==3.2.7
django-admin-thumbnails==0.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-mptt==0.13.3
mysqlclient==2.0.3
Pillow==8.3.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
